requirements.txt
python-dotenv
openai
pytz
boto3
python-telegram-bot==20.3